update: 		Controller

+++ src/main/java/net/zzh/sec/persistence/dao/IRoleJpaDAO.java
@@ -5,6 +5,6 @@ import net.zzh.sec.model.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface IRoleJpaDAO extends JpaRepository<Role, Long>, JpaSpecificationExecutor<Role>, INameSupport<Role> {
+public interface IRoleJpaDAO extends JpaRepository<Role, Integer>, JpaSpecificationExecutor<Role>, INameSupport<Role> {
     //
 }
