+++ src/main/java/net/zzh/sec/persistence/dao/IRolePermissionJpaDAO.java
@@ -5,6 +5,6 @@ import net.zzh.sec.model.RolePermission;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface IRolePermissionJpaDAO extends JpaRepository<RolePermission, Long>, JpaSpecificationExecutor<RolePermission>, INameSupport<RolePermission> {
+public interface IRolePermissionJpaDAO extends JpaRepository<RolePermission, Integer>, JpaSpecificationExecutor<RolePermission>, INameSupport<RolePermission> {
     //
 }
